[Chore] Separate out `vllm.utils.mem_utils` (#27143)
